Changed the folder structure ,Handled the errors properly and modified some api
