--- .env
@@ -1,8 +1,0 @@
-host     = "localhost"
-user     = "postgres"
-password = "password"
-dbname   = "blog_post_db"
-port     = "5432"
-http_port = ":5030"
-file_name = "logs.txt"
-secret_key = "da2sd33jas256o7dlji7s4ad"
