+++ api/routes/auth.go
@@ -1,0 +1,12 @@
+package routes
+
+import (
+	"blogs/api/handlers"
+
+	"github.com/labstack/echo/v4"
+)
+
+func AuthRoute(server *echo.Echo,handler *handlers.Handlers) {
+	server.POST("/signup", handler.Signup)
+	server.POST("/login", handler.Login)
+}
