+++ cmd/.env
@@ -1,0 +1,8 @@
+DB_HOST     = "localhost"
+DB_USER     = "postgres"
+DB_PASSWORD = "password"
+DB_NAME     = "blog_post_db"
+DB_PORT     = "5432"
+HTTP_PORT   = ":5030"
+FILE_NAME   = "logs.txt"
+SECRET_KEY  = "da2sd33jas256o7dlji7s4ad"
