+++ cmd/logs.txt
@@ -1,0 +1,4 @@
+INFO2024/12/02 20:38:47 connection.go:23: Connected to the server
+INFO2024/12/02 20:38:47 migration.go:17: Migrated tables successfully...
+INFO2024/12/02 20:39:22 connection.go:23: Connected to the server
+INFO2024/12/02 20:39:22 migration.go:17: Migrated tables successfully...
